clients: move some components to sailui
