+++ clients/bitwindow/lib/pages/sidechain_proposal_page.dart
@@ -1,7 +1,6 @@
 import 'package:auto_route/annotations.dart';
 import 'package:bitwindow/servers/api.dart';
 import 'package:bitwindow/widgets/qt_container.dart';
-import 'package:bitwindow/widgets/qt_icon_button.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:sail_ui/sail_ui.dart';
