+++ clients/bitwindow/lib/pages/sidechains_page.dart
@@ -3,12 +3,10 @@ import 'dart:math';
 import 'package:auto_route/auto_route.dart';
 import 'package:bitwindow/gen/wallet/v1/wallet.pbgrpc.dart';
 import 'package:bitwindow/pages/sidechain_activation_management_page.dart';
-import 'package:bitwindow/pages/wallet_page.dart';
 import 'package:bitwindow/providers/sidechain_provider.dart';
 import 'package:bitwindow/providers/transactions_provider.dart';
 import 'package:bitwindow/servers/api.dart';
 import 'package:bitwindow/widgets/error_container.dart';
-import 'package:bitwindow/widgets/qt_icon_button.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
