+++ clients/sail_ui/lib/sail_ui.dart
@@ -17,6 +17,7 @@ export './style/style.dart';
 export './theme/theme.dart';
 export './theme/theme_data.dart';
 export './widgets/buttons/button.dart';
+export './widgets/buttons/qt_icon_button.dart';
 export './widgets/buttons/toggle.dart';
 export './widgets/buttons/toggle_theme_button.dart';
 export './widgets/chips/action_header_chip.dart';
@@ -52,9 +53,11 @@ export './widgets/inputs/dropdown_menu.dart';
 export './widgets/inputs/large_embedded_input.dart';
 export './widgets/inputs/menu.dart';
 export './widgets/inputs/menu_items.dart';
+export './widgets/inputs/numeric_textfield.dart';
 export './widgets/inputs/radio_button.dart';
 export './widgets/inputs/text_field.dart';
 export './widgets/inputs/text_form_field.dart';
+export './widgets/inputs/unit_dropdown.dart';
 export './widgets/loading_indicator.dart';
 export './widgets/nav/nav_components.dart';
 export './widgets/optional_builder.dart';
